+++ commands/moderation/sanction.js
@@ -69,7 +69,7 @@ const getReactions = (msg,embed) => new Promise((resolve, reject) => {
           return;
         }
         msg.client.removeListener('messageReactionAdd', arguments.callee);
-        messageReaction.message.delete();
+        if(messageReaction.message) messageReaction.message.delete();
     
         return resolve({ response: messageReaction.emoji});
     }
